requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
biopython==1.84
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
numpy==2.1.3
openai==1.55.3
pandas==2.2.3
pydantic==2.10.2
pydantic_core==2.27.1
PyMuPDF==1.24.14
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv
